django_react_bioinfo/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
kombu==4.6.11
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
vine==1.3.0
